requirements-extra.txt
torch~=1.5.0
torchvision~=0.6.0
pytorch-mighty~=0.2.1
